# doxee-openapi.json
{"openapi":"3.0.0","paths":{"/api/v0/health-ckeck":{"get":{"operationId":"HealthCkeckController_check","parameters":[],"responses":{"200":{"description":""}}}},"/api/v0/secure/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"200":{"description":"User authenticated successfully."},"401":{"description":"Credentials are invalid."}},"tags":["auth"]}},"/api/v0/secure/auth/signUp":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpUserDto"}}}},"responses":{"200":{"description":"User created successfully."},"400":{"description":"Fields doesn't respect application constraints"},"500":{"description":"Error during registration process."}},"tags":["auth"]}},"/api/v0/secure/users":{"get":{"operationId":"UsersController_getUser","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/v0/secure/users/generateNewProposal":{"get":{"operationId":"UsersController_generateNewProposal","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProposalDto"}}}}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/v0/secure/users/acceptInvitation":{"post":{"operationId":"UsersController_acceptInvitation","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/v0/secure/users/createPeerNode":{"post":{"operationId":"UsersController_createPeerNode","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/v0/auth-code/all":{"get":{"operationId":"AuthCodeController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/api/v0/auth-code/valid":{"get":{"operationId":"AuthCodeController_findAllValid","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/api/v0/secure/chain-document/all":{"get":{"operationId":"ChainDocumentController_getAll","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"array"}}}}}},"tags":["chain-documents"],"security":[{"bearer":[]}]}},"/api/v0/secure/chain-document/{id}":{"get":{"operationId":"ChainDocumentController_getOneById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChainDocumentDto"}}}}},"tags":["chain-documents"],"security":[{"bearer":[]}]}},"/api/v0/secure/chain-document/upload":{"post":{"operationId":"ChainDocumentController_uploadChainDocument","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionDto"}}}}},"tags":["chain-documents"],"security":[{"bearer":[]}]}}},"info":{"title":"Doxee APIs","description":"Doxee APIs to upload files into Hyperledger Fabric v2 blockchain","version":"0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:8888"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginUserDto":{"properties":{"email":{"format":"email","type":"string","minLength":1},"password":{"type":"string"},"code":{"pattern":"^[-+]?[0-9]+$","type":"string","maxLength":8,"minLength":8}},"type":"object","required":["email","password"]},"SignUpUserDto":{"properties":{"name":{"pattern":"^[a-zA-Z][a-zA-Z]*$","type":"string","maxLength":32,"minLength":1},"surname":{"pattern":"^[a-zA-Z][a-zA-Z]*$","type":"string","maxLength":32,"minLength":1},"organization":{"pattern":"^(?!-|[0-9])(?!.*-$)(?!.*?--)[a-zA-Z0-9-]+$","type":"string","maxLength":64,"minLength":1},"awsClientId":{"pattern":"^[0-9]{12}$","type":"string","minLength":1},"email":{"format":"email","type":"string","minLength":1},"password":{"type":"string","minLength":1}},"type":"object","required":["name","surname","organization","awsClientId","email","password"]},"TransactionDto":{"properties":{"transactionId":{"type":"string"}},"type":"object","required":["transactionId"]},"ProposalDto":{"properties":{"proposalId":{"minLength":1,"type":"string"},"creationDate":{"oneOf":[{"format":"date","type":"string"},{"format":"date-time","type":"string"}],"minLength":1,"type":"string"},"expirationDate":{"oneOf":[{"format":"date","type":"string"},{"format":"date-time","type":"string"}],"minLength":1,"type":"string"},"status":{"enum":["IN_PROGRESS","APPROVED","REJECTED","EXPIRED","ACTION_FAILED"],"type":"string","minLength":1}},"type":"object","required":["proposalId","creationDate","expirationDate","status"]},"InvitationDto":{"properties":{"invitationId":{"minLength":1,"type":"string"},"creationDate":{"oneOf":[{"format":"date","type":"string"},{"format":"date-time","type":"string"}],"minLength":1,"type":"string"},"expirationDate":{"oneOf":[{"format":"date","type":"string"},{"format":"date-time","type":"string"}],"minLength":1,"type":"string"},"status":{"enum":["PENDING","ACCEPTING","ACCEPTED","REJECTED","EXPIRED"],"type":"string","minLength":1}},"type":"object","required":["invitationId","creationDate","expirationDate","status"]},"MemberDto":{"properties":{"memberId":{"minLength":1,"type":"string"},"name":{"minLength":1,"type":"string"},"creationDate":{"oneOf":[{"format":"date","type":"string"},{"format":"date-time","type":"string"}],"minLength":1,"type":"string"},"isOwned":{"type":"string","minLength":1},"status":{"enum":["CREATING","AVAILABLE","CREATE_FAILED","UPDATING","DELETING","DELETED","INACCESSIBLE_ENCRYPTION_KEY"],"type":"string","minLength":1}},"type":"object","required":["memberId","name","creationDate","isOwned","status"]},"NodeDto":{"properties":{"nodeId":{"minLength":1,"type":"string"},"creationDate":{"oneOf":[{"format":"date","type":"string"},{"format":"date-time","type":"string"}],"minLength":1,"type":"string"},"memberId":{"minLength":1,"type":"string"},"status":{"enum":["CREATING","AVAILABLE","UNHEALTHY","CREATE_FAILED","UPDATING","DELETING","DELETED","FAILED","INACCESSIBLE_ENCRYPTION_KEY"],"type":"string","minLength":1}},"type":"object","required":["nodeId","creationDate","memberId","status"]},"UserDto":{"properties":{"name":{"minLength":1,"type":"string"},"surname":{"minLength":1,"type":"string"},"organization":{"minLength":1,"type":"string"},"awsClientId":{"pattern":"^[0-9]{12}$","type":"string","minLength":1,"maxLength":12},"email":{"pattern":"^[A-Z0-9._%+-]+@[A-Z0-9.-]+\\.[A-Z]{2,}$","type":"string","minLength":1},"proposals":{"items":{},"type":"array"},"invitations":{"items":{},"type":"array"},"members":{"items":{},"type":"array"},"nodes":{"items":{},"type":"array"}},"type":"object","required":["name","surname","organization","awsClientId","email","proposals","invitations","members","nodes"]},"ChainDocumentDto":{"properties":{"id":{"minLength":1,"type":"string"},"name":{"minLength":1,"type":"string"},"uploadDate":{"oneOf":[{"format":"date","type":"string"},{"format":"date-time","type":"string"}],"minLength":1,"type":"string"}},"type":"object","required":["id","name","uploadDate"]}}}}